비밀번호 찾기 수정 및 리팩토링
테스트 완료.
