+++ backend/src/main/resources/application.properties
@@ -33,7 +33,7 @@ spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.Str
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
 spring.mail.username=publicmailbalsong@gmail.com
-spring.mail.password=
+spring.mail.password=tgkk ffxj xsxi xktv
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
 spring.mail.properties.mail.smtp.starttls.required=true
